be/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
alembic==1.13.1
python-dotenv==1.0.0
email-validator==2.1.0
fastapi-pagination==0.12.14
pandas==2.2.0
scikit-learn==1.4.0
cryptography==41.0.7
faker==22.0.0
